[RND-2329] psplash: Disable startup message
The psplash application enables a startup message by default, which creates
a white strip on the boot splash image as a background for the message.
Variscite's splash image already has a white background, so the white strip
goes unnoticed in this case.

Disable the startup message to avoid displaying a white strip on the screen.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>