var-create-yocto-sdcard: update for Scarthgap
This fixes the creation of an SD card.
After the Scarthgap update, the image names that are created by Yocto
have changed. Most specifically, while we previously used the symbolic
links as they had the correct name, we now want to use the regular files
because they have the correct names.

Also fix inconsistencies about what to specify on the command line. For
some parameters, "rootfs" had to be added to be able to run the script,
for some parameters it had to be NOT given.

This also eleminates the use of the `rename` tool, which is not always
present in identical versions on different distributions (Debian
installs perl script instead of coreutils version that requires
completely different parameters).

Signed-off-by: Matthias Proske <matthias.p@variscite.com>